LayoutTests/storage/indexeddb/modern/idbobjectstore-count-failures.html

This tests some obvious failures that can happen while calling IDBObjectStore.count().
<script>

if (window.testRunner) {
    testRunner.waitUntilDone();
    testRunner.dumpAsText();
}

function done()
{
    alert("Done");
    if (window.testRunner)
        testRunner.notifyDone();
}

var createRequest = window.indexedDB.open("IDBObjectStoreCountFailuresDatabase", 1);
var database;

createRequest.onupgradeneeded = function(event) {
    alert("Initial upgrade needed: Old version - " + event.oldVersion + " New version - " + event.newVersion);

    var versionTransaction = createRequest.transaction;
    database = event.target.result;
    var objectStore = database.createObjectStore("TestObjectStore");
    var request = objectStore.put("bar", "foo");

    request.onsuccess = function() {
        try {
            objectStore.count(NaN);
        } catch(e) {
            alert("Failed to count records in object store with an invalid key");
        }
        
        database.deleteObjectStore("TestObjectStore");
        
        try {
            objectStore.count();
        } catch(e) {
            alert("Failed to count records in object store that's been deleted");
        } 

        // Recreate the objectstore because we'll need it in phase 2.
        var objectStore = database.createObjectStore("TestObjectStore");
        objectStore.put("bar", "foo");    
    }
    
    versionTransaction.onabort = function(event) {
        alert("Initial upgrade versionchange transaction unexpected aborted");
        done();
    }

    versionTransaction.oncomplete = function(event) {
        alert("Initial upgrade versionchange transaction complete");
        continueTest1();
    }

    versionTransaction.onerror = function(event) {
        alert("Initial upgrade versionchange transaction unexpected error" + event);
        done();
    }
}

function continueTest1()
{
    var transaction = database.transaction("TestObjectStore", "readonly");
    var objectStore = transaction.objectStore("TestObjectStore");

    // Queue up a whole bunch of gets to keep the transaction alive for awhile
    for (var i = 0; i < 10; ++i)
        objectStore.get("foo");

    var getWhileInactive = function() {
        try {
            objectStore.count();
        } catch(e) {
            alert("Failed to count records in object store while transaction is inactive");
        } 
    }
    
    setTimeout(getWhileInactive, 0);
    
    transaction.onabort = function(event) {
        alert("readonly transaction unexpected abort" + event);
        done();
    }

    transaction.oncomplete = function(event) {
        alert("readonly transaction complete");
        done();
    }

    transaction.onerror = function(event) {
        alert("readonly transaction unexpected error" + event);
        done();
    }
}

</script>
